final updates

+++ src/Kanbas/Courses/Modules/index.tsx
@@ -2,16 +2,25 @@ import React, { useState, useEffect } from "react";
 import { BsGripVertical } from "react-icons/bs";
 import { useParams } from "react-router";
 import { useSelector, useDispatch } from "react-redux";
+import { RootState } from '../../store';
 import { setModules, addModule, editModule, updateModule, deleteModule } from "./reducer";
 import ModulesControls from "./ModulesControls";
 import LessonControlButtons from "./LessonControlButtons";
 import ModuleControlButtons from "./ModuleControlButtons";
 import * as client from "./client";
+
+interface User {
+  _id: string;
+  username: string;
+  role: string;
+}
+
 export default function Modules() {
   const { cid } = useParams();
   const [moduleName, setModuleName] = useState("");
   const modules = useSelector((state: any) => state.modulesReducer.modules);
   const dispatch = useDispatch();
+  const currentUser = useSelector((state: RootState) => state.accountReducer.currentUser as User | null);
   const fetchModules = async () => {
     const modules = await client.findModulesForCourse(cid as string);
     dispatch(setModules(modules));
@@ -21,12 +30,17 @@ export default function Modules() {
   }, []);
 
   const createModule = async (module: any) => {
+    
     const newModule = await client.createModule(cid as string, module);
     dispatch(addModule(newModule));
   };
   const removeModule = async (moduleId: string) => {
-    await client.deleteModule(moduleId);
-    dispatch(deleteModule(moduleId));
+    if (currentUser?.role === "FACULTY") {
+      await client.deleteModule(moduleId);
+      dispatch(deleteModule(moduleId));
+    } else {
+      alert("Only faculty can delete modules.");
+    }
   };
 
   const saveModule = async (module: any) => {
@@ -40,8 +54,12 @@ export default function Modules() {
         setModuleName={setModuleName}
         moduleName={moduleName}
         addModule={() => {
-          createModule({ name: moduleName, course: cid });
-          setModuleName("");
+          if (currentUser?.role === "FACULTY") {
+            createModule({ name: moduleName, course: cid });
+            setModuleName("");
+          } else {
+            alert("Only faculty can create modules.");
+          }
         }}
       />
       <br />
@@ -59,7 +77,7 @@ export default function Modules() {
               <div className="wd-title p-3 ps-2 bg-secondary">
                 <BsGripVertical className="me-2 fs-3" />
                 {!module.editing && module.name}
-                {module.editing && (
+                {module.editing && currentUser?.role === "FACULTY" && (
                   <input
                     className="form-control w-50 d-inline-block"
                     onChange={(e) => saveModule({ ...module, name: e.target.value })}
@@ -73,8 +91,20 @@ export default function Modules() {
                 )}
                 <ModuleControlButtons
                   moduleId={module._id}
-                  deleteModule={(moduleId) => { removeModule(moduleId); }}
-                  editModule={() => dispatch(editModule(module._id))}
+                  deleteModule={(moduleId) => { 
+                    if (currentUser?.role === "FACULTY") {
+                      removeModule(moduleId); 
+                    } else {
+                      alert("Only faculty can delete modules.");
+                    }
+                  }}
+                  editModule={() => {
+                    if (currentUser?.role === "FACULTY") {
+                      dispatch(editModule(module._id));
+                    } else {
+                      alert("Only faculty can edit modules.");
+                    }
+                  }}
                 />
               </div>
               {module.lessons && (
